Redesign CloudTrail watcher and auditor
This changes how the CloudTrail watcher functions so that it now keeps track
of all CloudTrails configured in each region and across all regions. Some of
this change is driven by the addition of the [AWS support for multiple
CloudTrails in each region](https://aws.amazon.com/blogs/aws/aws-cloudtrail-update-turn-on-in-all-regions-use-multiple-trails/).
This change enables the watcher to detect when a region is missing a
CloudTrail which writes to the Mozilla CloudTrail bucket and to detect when no
CloudTrail is configured to record GlobalServiceEvents like IAM events. The
watcher communicates these missing CloudTrails by creating fake/phantom items
in Security Monkey for these missing CloudTrails for the auditor to notice.
This is not an ideal method but I haven't come up with something better yet.